Create `hello-world` image
