+++ src/images.rs
@@ -2,6 +2,7 @@ pub mod coblox_bitcoincore;
 pub mod dynamodb_local;
 pub mod elasticmq;
 pub mod generic;
+pub mod hello_world;
 pub mod mongo;
 pub mod orientdb;
 pub mod parity_parity;
